Refactor all actions into separate files
