+++ src/engine/Actions/Action/Repeat.ts
@@ -1,43 +1,44 @@
-import { Actor } from '../Actor';
-import { Action } from './Action';
-import { ActionContext } from './ActionContext';
-import { ActionQueue } from './ActionQueue';
+import { Actor } from '../../Actor';
+import { Action } from '../Action';
+import { ActionContext } from '../ActionContext';
+import { ActionQueue } from '../ActionQueue';
 
-export class RepeatForever implements Action {
+export class Repeat implements Action {
   private _actionQueue: ActionQueue;
+  private _repeat: number;
+  private _originalRepeat: number;
   private _stopped: boolean = false;
   private _repeatContext: ActionContext;
   private _repeatBuilder: (repeatContext: ActionContext) => any;
-  constructor(actor: Actor, repeatBuilder: (repeatContext: ActionContext) => any) {
+  constructor(actor: Actor, repeatBuilder: (repeatContext: ActionContext) => any, repeat: number) {
     this._repeatBuilder = repeatBuilder;
     this._repeatContext = new ActionContext(actor);
     this._actionQueue = this._repeatContext.getQueue();
 
+    this._repeat = repeat;
+    this._originalRepeat = repeat;
+
     this._repeatBuilder(this._repeatContext);
+    this._repeat--; // current execution is the first repeat
   }
 
   public update(delta: number): void {
-    if (this._stopped) {
-      return;
-    }
-
     if (this._actionQueue.isComplete()) {
       this._actionQueue.reset();
+      this._repeat--;
     }
-
     this._actionQueue.update(delta);
   }
 
   public isComplete(): boolean {
-    return this._stopped;
+    return this._stopped || (this._repeat <= 0 && this._actionQueue.isComplete());
   }
 
   public stop(): void {
     this._stopped = true;
-    this._actionQueue.clearActions();
   }
 
   public reset(): void {
-    return;
+    this._repeat = this._originalRepeat;
   }
 }
