+++ src/engine/Actions/Action/RepeatForever.ts
@@ -1,44 +1,49 @@
-import { Actor } from '../Actor';
-import { Action } from './Action';
-import { ActionContext } from './ActionContext';
-import { ActionQueue } from './ActionQueue';
+import { Actor } from '../../Actor';
+import { Action } from '../Action';
+import { ActionContext } from '../ActionContext';
+import { ActionQueue } from '../ActionQueue';
 
-export class Repeat implements Action {
+/**
+ * RepeatForever Action implementation, it is recommended you use the fluent action
+ * context API.
+ *
+ *
+ */
+export class RepeatForever implements Action {
   private _actionQueue: ActionQueue;
-  private _repeat: number;
-  private _originalRepeat: number;
   private _stopped: boolean = false;
   private _repeatContext: ActionContext;
   private _repeatBuilder: (repeatContext: ActionContext) => any;
-  constructor(actor: Actor, repeatBuilder: (repeatContext: ActionContext) => any, repeat: number) {
+  constructor(actor: Actor, repeatBuilder: (repeatContext: ActionContext) => any) {
     this._repeatBuilder = repeatBuilder;
     this._repeatContext = new ActionContext(actor);
     this._actionQueue = this._repeatContext.getQueue();
 
-    this._repeat = repeat;
-    this._originalRepeat = repeat;
-
     this._repeatBuilder(this._repeatContext);
-    this._repeat--; // current execution is the first repeat
   }
 
   public update(delta: number): void {
+    if (this._stopped) {
+      return;
+    }
+
     if (this._actionQueue.isComplete()) {
       this._actionQueue.reset();
-      this._repeat--;
     }
+
     this._actionQueue.update(delta);
   }
 
   public isComplete(): boolean {
-    return this._stopped || (this._repeat <= 0 && this._actionQueue.isComplete());
+    return this._stopped;
   }
 
   public stop(): void {
     this._stopped = true;
+    this._actionQueue.clearActions();
   }
 
   public reset(): void {
-    this._repeat = this._originalRepeat;
+    return;
   }
 }
